requirements.txt
PyQt5>=5.15.0
torch>=2.0.0
torchaudio>=2.0.0
torchcodec
openai-whisper>=20231117
qwen-asr>=0.0.6
silero-vad>=5.0.0
funasr>=1.0.0
noisereduce>=3.0.0
scipy>=1.11.0
soundfile>=0.12.0
numpy>=1.24.0
loguru>=0.7.0
ffmpeg-python>=0.2.0
scikit-learn>=1.3.0
